requirements.txt
streamlit>=1.37
python-docx>=1.1.2
openai>=1.40.0

